Bokeh Dependency and Taxes Bar Chart
Added Bokeh 0.12.15 as a dependency and implemented the first simple bar
chart using Bokeh - a bar graph of income over time. Currently the API
is not being used as a data source for this graph, instead I'm just
embedding HTML and JavaScript directly into the taxes page, but it's a
good first start!