Accounts app and models
The accounts app is the primary financial tracking application for
ledger. It maintains a relationship between financial accounts of
varying types (e.g. cash, credit, investments, billing, etc) to a
monthly balance sheet that has balances and transactions. The montly
timeseries data per balance sheet is the primary target of tracking.

The models in the accounts app are complex and interrelated, but I think
this commit shows the best possible ER diagram for the model.