Inline editing of balances and transactions
Closes #7

Added inline editing of balances and transactions in the admin view, as
well as filtered the fields displayed for easy editing. Also added a
signal so that when the balance sheet or balance are saved the ending
balance is also updated. Finally, removed future restriction from
balance sheet view so that future balance sheets are displayed.