Adds fixed account balance width and accounting format
Sets the sidebar display to none on XS (iphone screen) to prevent screen
overlap bug with the sidebar.

Implements accounting template tags to display balances and amounts on
the balance sheet screen with a currency and aligned on the decimal.
Negative values are displayed in parens instead of with a negative sign.

Uses .col classes to fix the width of column tables in both the accounts
tables and the transactions table.

Fixes #10
Resolves #4
Resolves #3