Added balance sheet print button and view
Added a button to the balance sheet view that appends a ?print=true
query to the url, then modified the view to return a print template in
this case instead of the dashboard layout. If background graphics are
included in the printing, it actually prints a very nice sheet to PDF.

Note that the print button is a snippet that can be used elsewhere.

Resolves #9