Added account totals at the bottom of account types table
Modified the balance sheet templates to create an accounts table and
transactions table template stub for reusability. Added a queryset that
does a better job of collecting account balances, and added an
aggregation to compute the total (also a migration to exclude accounts
from computation while still allowing them to be active).

Resolves #11