Calendar icon archive and next sheet button
Modifies the balance sheets archive to use a CSS calendar button to
display the archive with quick links to each.

Adds a next_sheet button which allows a quick link if the latest sheet
is with 15 days of the current date, or a create next balance sheet if
it is not and it doesn't exist.

Closes #5