Added tests runners and factories
This commit adds scaffolding for creating tests for the ledger app
including setting up pytest and pytest-django for test running and
flakes and coverage for detailed test information. Factory boy was
also installed for the creation of fixtures and some factories
created. Extensive tests have not been developed but this is a good
start to the testing process for this app.

Fixes #13