Accounting tags change None to 0
When there are no accounts or balance summations the django template
system returns None which caused an exception to be raised when doing
math with none. Now the template tags interpret None as 0 which means
the template does not raise an exception. This might not be the best way
to handle this, but it is a quick way to fix the problem.

Fixes #33